+++ lib/utils/RedisCommander.ts
@@ -3871,6 +3871,23 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
   ): Result<number, Context>;
 
   /**
+   * Set expiry for hash field using relative time to expire (seconds)
+   * - _group_: hash
+   * - _complexity_: O(N) where N is the number of specified fields
+   * - _since_: 7.4.0
+   */
+  hexpire(...args: [key: RedisKey, seconds: number | string, fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, nx: 'NX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, nx: 'NX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, xx: 'XX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, xx: 'XX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, gt: 'GT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, gt: 'GT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, lt: 'LT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hexpire(...args: [key: RedisKey, seconds: number | string, lt: 'LT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+
+  /**
    * Get the value of a hash field
    * - _group_: hash
    * - _complexity_: O(1)
@@ -4010,6 +4027,23 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
   hmset(
     ...args: [key: RedisKey, ...fieldValues: (string | Buffer | number)[]]
   ): Result<"OK", Context>;
+
+  /**
+   * Set expiry for hash field using relative time to expire (milliseconds)
+   * - _group_: hash
+   * - _complexity_: O(N) where N is the number of specified fields
+   * - _since_: 7.4.0
+   */
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, nx: 'NX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, nx: 'NX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, xx: 'XX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, xx: 'XX', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, gt: 'GT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, gt: 'GT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, lt: 'LT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[], callback: Callback<number[]>]): Result<number[], Context>;
+  hpexpire(...args: [key: RedisKey, milliseconds: number | string, lt: 'LT', fieldsToken: 'FIELDS', numfields: number | string, ...fields: (string | Buffer)[]]): Result<number[], Context>;
 
   /**
    * Get one or multiple random fields from a hash
@@ -9469,6 +9503,87 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
   ): Result<number, Context>;
 
   /**
+   * Deletes one or multiple entries from the stream.
+   * - _group_: stream
+   * - _complexity_: O(1) for each single item to delete in the stream, regardless of the stream size.
+   * - _since_: 8.2.0
+   */
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[],
+      callback: Callback<unknown>
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[]
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      keepref: "KEEPREF",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[],
+      callback: Callback<unknown>
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      keepref: "KEEPREF",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[]
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      delref: "DELREF",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[],
+      callback: Callback<unknown>
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      delref: "DELREF",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[]
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      acked: "ACKED",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[],
+      callback: Callback<unknown>
+    ]
+  ): Result<unknown, Context>;
+  xdelex(
+    ...args: [
+      key: RedisKey,
+      acked: "ACKED",
+      idsToken: "IDS",
+      numids: number | string,
+      ...ids: (string | Buffer | number)[]
+    ]
+  ): Result<unknown, Context>;
+
+  /**
    * Create a consumer group.
    * - _group_: stream
    * - _complexity_: O(1)
@@ -10260,7 +10375,7 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
   ): Result<unknown, Context>;
 
   /**
-   * Trims the stream to (approximately if '~' is passed) a certain size
+   * Deletes messages from the beginning of a stream.
    * - _group_: stream
    * - _complexity_: O(N), with N being the number of evicted entries. Constant times are very small however, since entries are organized in macro nodes containing multiple entries that can be released with a single deallocation.
    * - _since_: 5.0.0
@@ -10275,8 +10390,56 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     key: RedisKey,
     maxlen: "MAXLEN",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;
   xtrim(
@@ -10291,8 +10454,62 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     maxlen: "MAXLEN",
     equal: "=",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;
   xtrim(
@@ -10307,8 +10524,62 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     maxlen: "MAXLEN",
     approximately: "~",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    maxlen: "MAXLEN",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;
   xtrim(
@@ -10321,8 +10592,56 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     key: RedisKey,
     minid: "MINID",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;
   xtrim(
@@ -10337,8 +10656,62 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     minid: "MINID",
     equal: "=",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    equal: "=",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;
   xtrim(
@@ -10353,8 +10726,62 @@ interface RedisCommander<Context extends ClientContext = { type: "default" }> {
     minid: "MINID",
     approximately: "~",
     threshold: string | Buffer | number,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    acked: "ACKED",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
     countToken: "LIMIT",
     count: number | string,
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    keepref: "KEEPREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    delref: "DELREF",
+    callback?: Callback<number>
+  ): Result<number, Context>;
+  xtrim(
+    key: RedisKey,
+    minid: "MINID",
+    approximately: "~",
+    threshold: string | Buffer | number,
+    countToken: "LIMIT",
+    count: number | string,
+    acked: "ACKED",
     callback?: Callback<number>
   ): Result<number, Context>;